Interest: Enable interest days to be passed via constructor (#57)

+++ midas-core/src/main/java/de/nihas101/midas/core/interest/row/DefaultInterestCalculationRow.java
@@ -57,7 +57,7 @@ public class DefaultInterestCalculationRow implements InterestCalculationRow {
     }
 
     @Override
-    public Integer interestDaysCount() {
+    public BigDecimal interestDaysCount() {
         return interestCalculationRow.interestDaysCount();
     }
